INC-54 refactor: change Sidebar story
